@@ -9,17 +9,18 @@
     "url": "https://github.com/derekfinlinson/powerapps-tools"
   },
   "devDependencies": {
-    "@types/jest": "^27.5.2",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
-    "beachball": "^2.43.1",
-    "eslint": "^7.32.0",
-    "jest": "^27.5.1",
-    "prettier": "^2.8.8",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^20.17.30",
+    "@typescript-eslint/eslint-plugin": "^8.29.1",
+    "@typescript-eslint/parser": "^8.29.1",
+    "beachball": "^2.52.0",
+    "eslint": "^9.24.0",
+    "jest": "^29.7.0",
+    "prettier": "^3.5.3",
     "rimraf": "^6.0.1",
-    "ts-jest": "^27.1.5",
+    "ts-jest": "^29.3.1",
     "ts-node": "^10.9.2",
-    "typescript": "^4.9.5"
+    "typescript": "^5.8.3"
   },
   "scripts": {
     "build": "pnpm -r --stream clean && pnpm -r --stream build",
@@ -33,7 +34,7 @@
     "beachball": "beachball"
   },
   "volta": {
-    "node": "18.20.2",
-    "pnpm": "9.12.3"
+    "node": "20.19.0",
+    "pnpm": "10.8.0"
   }
 }
