@@ -3,13 +3,14 @@
   "version": "4.0.1",
   "license": "MIT",
   "description": "Utilities for interacting with Dataverse environments",
-  "main": "lib/index.js",
+  "exports": "./lib/index.js",
   "bin": {
     "dataverse-utils": "lib/index.js"
   },
   "engines": {
-    "node": ">=18.0.0"
+    "node": ">=20.0.0"
   },
+  "type": "module",
   "files": [
     "/lib"
   ],
@@ -24,19 +25,18 @@
   },
   "dependencies": {
     "@azure/msal-node": "^2.10.0",
-    "commander": "^12.0.0",
-    "cryptr": "^6.2.0",
+    "commander": "^13.1.0",
+    "cryptr": "^6.3.0",
     "dataverse-webapi": "workspace:^2.6.0",
-    "figures": "^3.2.0",
-    "glob": "^10.3.10",
+    "figures": "^6.1.0",
+    "glob": "^11.0.1",
     "kleur": "^4.1.5",
     "prompts": "^2.4.2"
   },
   "devDependencies": {
-    "@types/cryptr": "^4.0.1",
+    "@types/cryptr": "^4.0.3",
     "@types/glob": "^8.1.0",
-    "@types/node": "^18.19.20",
-    "@types/prompts": "^2.4.4"
+    "@types/prompts": "^2.4.9"
   },
   "volta": {
     "extends": "../../package.json"
