@@ -1,7 +1,7 @@
 import prompts from 'prompts';
-import { cacheExists, cachePlugin, deleteCache } from './cachePlugin';
+import { cacheExists, cachePlugin, deleteCache } from './cachePlugin.js';
 import { AuthenticationResult, PublicClientApplication } from '@azure/msal-node';
-import { logger } from './logger';
+import { logger } from './logger.js';
 
 const clientId = '51f81489-12ee-4a9e-aaae-a2591f45987d';
 
