@@ -1,8 +1,8 @@
 #!/usr/bin/env node
 import { program } from 'commander';
 
-import deploy from './deploy';
-import generate from './generate';
+import deploy from './deploy.js';
+import generate from './generate.js';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const packageJson = require('../package.json');
